+++ swanlab/data/backup/__init__.py
@@ -5,7 +5,7 @@
 @description: 日志备份处理模块
 """
 
-from .handler import BackupHandler, backup, DataStore
+from .handler import BackupHandler, DataStore
 
 
-__all__ = ["BackupHandler", "backup", "DataStore"]
+__all__ = ["BackupHandler", "DataStore"]
