+++ swanlab/data/callbacker/offline.py
@@ -7,17 +7,18 @@
 
 from rich.text import Text
 
-from swanlab.data.backup import backup
 from swanlab.data.run.callback import SwanLabRunCallback
 from swanlab.log import swanlog
 from swanlab.log.type import LogData
 from swanlab.toolkit import ColumnInfo, MetricInfo, RuntimeInfo
+from ..backup import BackupHandler
 from ..run import get_run
 
 
 class OfflineCallback(SwanLabRunCallback):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
+    def __init__(self):
+        super().__init__()
+        self.device = BackupHandler()
 
     def __str__(self) -> str:
         return "SwanLabOfflineCallback"
@@ -34,7 +35,6 @@ class OfflineCallback(SwanLabRunCallback):
             sep="",
         )
 
-    @backup("terminal")
     def _terminal_handler(self, log_data: LogData):
         """
         终端输出写入操作
@@ -45,9 +45,9 @@ class OfflineCallback(SwanLabRunCallback):
 
     def on_init(self, proj_name: str, workspace: str, public: bool = None, logdir: str = None, *args, **kwargs):
         # 设置项目缓存
-        self.backup.cache_proj_name = proj_name
-        self.backup.cache_workspace = workspace
-        self.backup.cache_public = public
+        self.device.cache_proj_name = proj_name
+        self.device.cache_workspace = workspace
+        self.device.cache_public = public
 
     def on_run(self, *args, **kwargs):
         self.handle_run()
@@ -55,19 +55,16 @@ class OfflineCallback(SwanLabRunCallback):
         swanlog.info("Backing up run", Text(self.settings.exp_name, "yellow"), "locally")
         self._sync_tip_print()
 
-    @backup('runtime')
     def on_runtime_info_update(self, r: RuntimeInfo, *args, **kwargs):
         pass
 
-    @backup("column")
     def on_column_create(self, column_info: ColumnInfo, *args, **kwargs):
         pass
 
-    @backup("metric")
     def on_metric_create(self, metric_info: MetricInfo, *args, **kwargs):
         pass
 
     def on_stop(self, error: str = None, *args, **kwargs):
         self._sync_tip_print()
-        self.backup.stop(error=error, epoch=get_run().swanlog_epoch + 1)
+        self.device.stop(error=error, epoch=get_run().swanlog_epoch + 1)
         self._unregister_sys_callback()
